Step 10: ruff clean across repo, full suite green
Remove a redundant int() cast in the service test helper, apply ruff's import
sorting and formatting across the touched files. Final state: 53 tests pass,
ruff check + format clean.

+++ tests/test_services.py
@@ -63,7 +63,7 @@ def _make_entry(coord: ValloxDataUpdateCoordinator) -> MagicMock:
 
 
 def _set_raw(raw_map: dict[str, int | None], key: str, celsius: float) -> None:
-    raw_map[key] = int(round((celsius + 273.15) * 100))
+    raw_map[key] = round((celsius + 273.15) * 100)
 
 
 def _patch_read_raw(monkeypatch, raw_map: dict[str, int | None]) -> None:
